requirements.txt
certifi==2019.6.16
Django==2.2.3
django-phone-field==1.7.1
pipenv==2018.11.26
pytz==2019.1
sqlparse==0.3.0
virtualenv==16.6.2
virtualenv-clone==0.5.3
